ENH: clean embedding model cache when using vllm engine (#3956)
